fix: exiftool镜像地址失效时使用原地址

+++ install.sh
@@ -5,18 +5,28 @@ if [ -f 'inited' ]; then
   exit 0
 fi
 
-# 下载文件
-curl -O http://file.lsvm.xyz/Image-ExifTool-12.92.tar.gz
+download_file_name='Image-ExifTool-12.92.tar.gz'
 
+# 优先镜像链接下载
+curl -O http://file.lsvm.xyz/$download_file_name
+
+if [ ! -f $download_file_name ]; then
+  curl -O https://exiftool.org/$download_file_name
+fi
+
+if [ ! -f $download_file_name ]; then
+  echo '下载失败,请检查网络连接状态或下载链接是否有效'
+  exit 1
+fi
 
 # 创建目录
 mkdir -p ./exiftool
 
 # 解压文件
-tar -xzf Image-ExifTool-12.92.tar.gz -C ./exiftool --strip-components=1
+tar -xzf $download_file_name -C ./exiftool --strip-components=1
 
 # 删除压缩包
-rm Image-ExifTool-12.92.tar.gz
+rm $download_file_name
 
 # 下载 python 依赖
 pip3 install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
